third_party/blink/web_tests/http/tests/security/cross-frame-access-put-expected.txt

ALERT: PASS: window.Attr should be 'function Attr() { [native code] }' and is.
ALERT: PASS: window.CDATASection should be 'function CDATASection() { [native code] }' and is.
ALERT: PASS: window.CharacterData should be 'function CharacterData() { [native code] }' and is.
ALERT: PASS: window.Comment should be 'function Comment() { [native code] }' and is.
ALERT: PASS: window.CSSRule should be 'function CSSRule() { [native code] }' and is.
ALERT: PASS: window.CSSStyleDeclaration should be 'function CSSStyleDeclaration() { [native code] }' and is.
ALERT: PASS: window.Document should be 'function Document() { [native code] }' and is.
ALERT: PASS: window.DocumentFragment should be 'function DocumentFragment() { [native code] }' and is.
ALERT: PASS: window.DocumentType should be 'function DocumentType() { [native code] }' and is.
ALERT: PASS: window.DOMException should be 'function DOMException() { [native code] }' and is.
ALERT: PASS: window.DOMImplementation should be 'function DOMImplementation() { [native code] }' and is.
ALERT: PASS: window.DOMParser should be 'function DOMParser() { [native code] }' and is.
ALERT: PASS: window.Element should be 'function Element() { [native code] }' and is.
ALERT: PASS: window.Entity should be 'undefined' and is.
ALERT: PASS: window.EntityReference should be 'undefined' and is.
ALERT: PASS: window.EvalError should be 'function EvalError() { [native code] }' and is.
ALERT: PASS: window.Event should be 'function Event() { [native code] }' and is.
ALERT: PASS: window.HTMLAnchorElement should be 'function HTMLAnchorElement() { [native code] }' and is.
ALERT: PASS: window.HTMLAreaElement should be 'function HTMLAreaElement() { [native code] }' and is.
ALERT: PASS: window.HTMLBaseElement should be 'function HTMLBaseElement() { [native code] }' and is.
ALERT: PASS: window.HTMLBodyElement should be 'function HTMLBodyElement() { [native code] }' and is.
ALERT: PASS: window.HTMLBRElement should be 'function HTMLBRElement() { [native code] }' and is.
ALERT: PASS: window.HTMLButtonElement should be 'function HTMLButtonElement() { [native code] }' and is.
ALERT: PASS: window.HTMLCanvasElement should be 'function HTMLCanvasElement() { [native code] }' and is.
ALERT: PASS: window.HTMLDirectoryElement should be 'function HTMLDirectoryElement() { [native code] }' and is.
ALERT: PASS: window.HTMLDivElement should be 'function HTMLDivElement() { [native code] }' and is.
ALERT: PASS: window.HTMLDListElement should be 'function HTMLDListElement() { [native code] }' and is.
ALERT: PASS: window.HTMLDocument should be 'function HTMLDocument() { [native code] }' and is.
ALERT: PASS: window.HTMLElement should be 'function HTMLElement() { [native code] }' and is.
ALERT: PASS: window.HTMLFieldSetElement should be 'function HTMLFieldSetElement() { [native code] }' and is.
ALERT: PASS: window.HTMLFontElement should be 'function HTMLFontElement() { [native code] }' and is.
ALERT: PASS: window.HTMLFormElement should be 'function HTMLFormElement() { [native code] }' and is.
ALERT: PASS: window.HTMLFrameElement should be 'function HTMLFrameElement() { [native code] }' and is.
ALERT: PASS: window.HTMLFrameSetElement should be 'function HTMLFrameSetElement() { [native code] }' and is.
ALERT: PASS: window.HTMLHeadElement should be 'function HTMLHeadElement() { [native code] }' and is.
ALERT: PASS: window.HTMLHeadingElement should be 'function HTMLHeadingElement() { [native code] }' and is.
ALERT: PASS: window.HTMLHRElement should be 'function HTMLHRElement() { [native code] }' and is.
ALERT: PASS: window.HTMLHtmlElement should be 'function HTMLHtmlElement() { [native code] }' and is.
ALERT: PASS: window.HTMLIFrameElement should be 'function HTMLIFrameElement() { [native code] }' and is.
ALERT: PASS: window.HTMLImageElement should be 'function HTMLImageElement() { [native code] }' and is.
ALERT: PASS: window.HTMLInputElement should be 'function HTMLInputElement() { [native code] }' and is.
ALERT: PASS: window.HTMLIsIndexElement should be 'undefined' and is.
ALERT: PASS: window.HTMLLabelElement should be 'function HTMLLabelElement() { [native code] }' and is.
ALERT: PASS: window.HTMLLegendElement should be 'function HTMLLegendElement() { [native code] }' and is.
ALERT: PASS: window.HTMLLIElement should be 'function HTMLLIElement() { [native code] }' and is.
ALERT: PASS: window.HTMLLinkElement should be 'function HTMLLinkElement() { [native code] }' and is.
ALERT: PASS: window.HTMLMapElement should be 'function HTMLMapElement() { [native code] }' and is.
ALERT: PASS: window.HTMLMarqueeElement should be 'function HTMLMarqueeElement() { [native code] }' and is.
ALERT: PASS: window.HTMLMenuElement should be 'function HTMLMenuElement() { [native code] }' and is.
ALERT: PASS: window.HTMLMetaElement should be 'function HTMLMetaElement() { [native code] }' and is.
ALERT: PASS: window.HTMLModElement should be 'function HTMLModElement() { [native code] }' and is.
ALERT: PASS: window.HTMLOListElement should be 'function HTMLOListElement() { [native code] }' and is.
ALERT: PASS: window.HTMLOptGroupElement should be 'function HTMLOptGroupElement() { [native code] }' and is.
ALERT: PASS: window.HTMLOptionElement should be 'function HTMLOptionElement() { [native code] }' and is.
ALERT: PASS: window.HTMLParagraphElement should be 'function HTMLParagraphElement() { [native code] }' and is.
ALERT: PASS: window.HTMLParamElement should be 'function HTMLParamElement() { [native code] }' and is.
ALERT: PASS: window.HTMLPreElement should be 'function HTMLPreElement() { [native code] }' and is.
ALERT: PASS: window.HTMLQuoteElement should be 'function HTMLQuoteElement() { [native code] }' and is.
ALERT: PASS: window.HTMLScriptElement should be 'function HTMLScriptElement() { [native code] }' and is.
ALERT: PASS: window.HTMLSelectElement should be 'function HTMLSelectElement() { [native code] }' and is.
ALERT: PASS: window.HTMLStyleElement should be 'function HTMLStyleElement() { [native code] }' and is.
ALERT: PASS: window.HTMLTableCaptionElement should be 'function HTMLTableCaptionElement() { [native code] }' and is.
ALERT: PASS: window.HTMLTableCellElement should be 'function HTMLTableCellElement() { [native code] }' and is.
ALERT: PASS: window.HTMLTableColElement should be 'function HTMLTableColElement() { [native code] }' and is.
ALERT: PASS: window.HTMLTableElement should be 'function HTMLTableElement() { [native code] }' and is.
ALERT: PASS: window.HTMLTableRowElement should be 'function HTMLTableRowElement() { [native code] }' and is.
ALERT: PASS: window.HTMLTableSectionElement should be 'function HTMLTableSectionElement() { [native code] }' and is.
ALERT: PASS: window.HTMLTextAreaElement should be 'function HTMLTextAreaElement() { [native code] }' and is.
ALERT: PASS: window.HTMLTitleElement should be 'function HTMLTitleElement() { [native code] }' and is.
ALERT: PASS: window.HTMLUListElement should be 'function HTMLUListElement() { [native code] }' and is.
ALERT: PASS: window.MutationEvent should be 'function MutationEvent() { [native code] }' and is.
ALERT: PASS: window.Node should be 'function Node() { [native code] }' and is.
ALERT: PASS: window.NodeFilter should be 'function NodeFilter() { [native code] }' and is.
ALERT: PASS: window.ProcessingInstruction should be 'function ProcessingInstruction() { [native code] }' and is.
ALERT: PASS: window.Range should be 'function Range() { [native code] }' and is.
ALERT: PASS: window.RangeError should be 'function RangeError() { [native code] }' and is.
ALERT: PASS: window.ReferenceError should be 'function ReferenceError() { [native code] }' and is.
ALERT: PASS: window.SyntaxError should be 'function SyntaxError() { [native code] }' and is.
ALERT: PASS: window.Text should be 'function Text() { [native code] }' and is.
ALERT: PASS: window.TypeError should be 'function TypeError() { [native code] }' and is.
ALERT: PASS: window.URIError should be 'function URIError() { [native code] }' and is.
ALERT: PASS: window.XMLDocument should be 'function XMLDocument() { [native code] }' and is.
ALERT: PASS: window.XMLSerializer should be 'function XMLSerializer() { [native code] }' and is.
ALERT: PASS: window.XPathEvaluator should be 'function XPathEvaluator() { [native code] }' and is.
ALERT: PASS: window.XPathResult should be 'function XPathResult() { [native code] }' and is.
ALERT: PASS: window.clientInformation should be '[object Navigator]' and is.
ALERT: PASS: window.closed should be 'false' and is.
ALERT: PASS: window.console should be '[object Object]' and is.
ALERT: PASS: window.crypto should be '[object Crypto]' and is.
ALERT: PASS: window.defaultStatus should be '' and is.
ALERT: PASS: window.defaultstatus should be '' and is.
ALERT: PASS: window.devicePixelRatio should be '1' and is.
ALERT: PASS: window.document should be '[object HTMLDocument]' and is.
ALERT: PASS: window.embeds should be 'undefined' and is.
ALERT: PASS: window.frameElement should be 'null' and is.
ALERT: PASS: window.frames should be '[object Window]' and is.
ALERT: PASS: window.history should be '[object History]' and is.
ALERT: PASS: window.images should be 'undefined' and is.
ALERT: PASS: window.innerHeight should be '150' and is.
ALERT: PASS: window.innerWidth should be '300' and is.
ALERT: PASS: window.length should be '0' and is.
ALERT: PASS: window.locationbar should be '[object BarProp]' and is.
ALERT: PASS: window.menubar should be '[object BarProp]' and is.
ALERT: PASS: window.name should be '' and is.
ALERT: PASS: window.navigator should be '[object Navigator]' and is.
ALERT: PASS: window.offscreenBuffering should be 'true' and is.
ALERT: PASS: window.onabort should be 'null' and is.
ALERT: PASS: window.onbeforeunload should be 'null' and is.
ALERT: PASS: window.onblur should be 'null' and is.
ALERT: PASS: window.onchange should be 'null' and is.
ALERT: PASS: window.onclick should be 'null' and is.
ALERT: PASS: window.ondblclick should be 'null' and is.
ALERT: PASS: window.onerror should be 'null' and is.
ALERT: PASS: window.onfocus should be 'null' and is.
ALERT: PASS: window.onkeydown should be 'null' and is.
ALERT: PASS: window.onkeypress should be 'null' and is.
ALERT: PASS: window.onkeyup should be 'null' and is.
ALERT: PASS: window.onload should be 'null' and is.
ALERT: PASS: window.onmousedown should be 'null' and is.
ALERT: PASS: window.onmousemove should be 'null' and is.
ALERT: PASS: window.onmouseout should be 'null' and is.
ALERT: PASS: window.onmouseover should be 'null' and is.
ALERT: PASS: window.onmouseup should be 'null' and is.
ALERT: PASS: window.onmousewheel should be 'null' and is.
ALERT: PASS: window.onreset should be 'null' and is.
ALERT: PASS: window.onresize should be 'null' and is.
ALERT: PASS: window.onscroll should be 'null' and is.
ALERT: PASS: window.onsearch should be 'null' and is.
ALERT: PASS: window.onselect should be 'null' and is.
ALERT: PASS: window.onsubmit should be 'null' and is.
ALERT: PASS: window.onunload should be 'null' and is.
ALERT: PASS: window.opener should be 'null' and is.
ALERT: PASS: window.outerHeight matched the expected value.
ALERT: PASS: window.outerWidth matched the expected value.
ALERT: PASS: window.pageXOffset should be '0' and is.
ALERT: PASS: window.pageYOffset should be '0' and is.
ALERT: PASS: window.personalbar should be '[object BarProp]' and is.
ALERT: PASS: window.plugins should be 'undefined' and is.
ALERT: PASS: window.screen should be '[object Screen]' and is.
ALERT: PASS: window.screenLeft matched the expected value.
ALERT: PASS: window.screenTop matched the expected value.
ALERT: PASS: window.screenX matched the expected value.
ALERT: PASS: window.screenY matched the expected value.
ALERT: PASS: window.scrollbars should be '[object BarProp]' and is.
ALERT: PASS: window.scrollX should be '0' and is.
ALERT: PASS: window.scrollY should be '0' and is.
ALERT: PASS: window.self should be '[object Window]' and is.
ALERT: PASS: window.status should be '' and is.
ALERT: PASS: window.statusbar should be '[object BarProp]' and is.
ALERT: PASS: window.toolbar should be '[object BarProp]' and is.
ALERT: PASS: window.window should be '[object Window]' and is.
ALERT: PASS: window.parent should be parentOld and is.
ALERT: PASS: window.top should be topOld and is.
ALERT: PASS: window.addEventListener should be 'function addEventListener() { [native code] }' and is.
ALERT: PASS: window.alert should be 'function alert() { [native code] }' and is.
ALERT: PASS: window.atob should be 'function atob() { [native code] }' and is.
ALERT: PASS: window.blur should be 'function blur() { [native code] }' and is.
ALERT: PASS: window.btoa should be 'function btoa() { [native code] }' and is.
ALERT: PASS: window.captureEvents should be 'function captureEvents() { [native code] }' and is.
ALERT: PASS: window.clearInterval should be 'function clearInterval() { [native code] }' and is.
ALERT: PASS: window.clearTimeout should be 'function clearTimeout() { [native code] }' and is.
ALERT: PASS: window.close should be 'function close() { [native code] }' and is.
ALERT: PASS: window.confirm should be 'function confirm() { [native code] }' and is.
ALERT: PASS: window.eval should be 'function eval() { [native code] }' and is.
ALERT: PASS: window.find should be 'function find() { [native code] }' and is.
ALERT: PASS: window.focus should be 'function focus() { [native code] }' and is.
ALERT: PASS: window.getComputedStyle should be 'function getComputedStyle() { [native code] }' and is.
ALERT: PASS: window.getMatchedCSSRules should be 'undefined' and is.
ALERT: PASS: window.getSelection should be 'function getSelection() { [native code] }' and is.
ALERT: PASS: window.moveBy should be 'function moveBy() { [native code] }' and is.
ALERT: PASS: window.moveTo should be 'function moveTo() { [native code] }' and is.
ALERT: PASS: window.open should be 'function open() { [native code] }' and is.
ALERT: PASS: window.print should be 'function print() { [native code] }' and is.
ALERT: PASS: window.prompt should be 'function prompt() { [native code] }' and is.
ALERT: PASS: window.releaseEvents should be 'function releaseEvents() { [native code] }' and is.
ALERT: PASS: window.removeEventListener should be 'function removeEventListener() { [native code] }' and is.
ALERT: PASS: window.resizeBy should be 'function resizeBy() { [native code] }' and is.
ALERT: PASS: window.resizeTo should be 'function resizeTo() { [native code] }' and is.
ALERT: PASS: window.scroll should be 'function scroll() { [native code] }' and is.
ALERT: PASS: window.scrollBy should be 'function scrollBy() { [native code] }' and is.
ALERT: PASS: window.scrollTo should be 'function scrollTo() { [native code] }' and is.
ALERT: PASS: window.setInterval should be 'function setInterval() { [native code] }' and is.
ALERT: PASS: window.setTimeout should be 'function setTimeout() { [native code] }' and is.
ALERT: PASS: window.stop should be 'function stop() { [native code] }' and is.
This test checks cross-frame access security of window attribute setters (rdar://problem/5326791).

PASS: Unable to set property Attr: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property CDATASection: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property CharacterData: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property Comment: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property CSSRule: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property CSSStyleDeclaration: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property Document: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property DocumentFragment: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property DocumentType: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property DOMException: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property DOMImplementation: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property DOMParser: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property Element: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property EvalError: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property Event: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLAnchorElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLAreaElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLBaseElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLBodyElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLBRElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLButtonElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLCanvasElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLDirectoryElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLDivElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLDListElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLDocument: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLFieldSetElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLFontElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLFormElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLFrameElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLFrameSetElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLHeadElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLHeadingElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLHRElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLHtmlElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLIFrameElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLImageElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLInputElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLLabelElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLLegendElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLLIElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLLinkElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLMapElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLMarqueeElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLMenuElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLMetaElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLModElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLOListElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLOptGroupElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLOptionElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLParagraphElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLParamElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLPreElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLQuoteElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLScriptElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLSelectElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLStyleElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLTableCaptionElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLTableCellElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLTableColElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLTableElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLTableRowElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLTableSectionElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLTextAreaElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLTitleElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property HTMLUListElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property MutationEvent: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property Node: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property NodeFilter: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property ProcessingInstruction: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property Range: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property RangeError: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property ReferenceError: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property SyntaxError: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property Text: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property TypeError: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property URIError: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property XMLDocument: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property XMLSerializer: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property XPathEvaluator: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property XPathResult: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property clientInformation: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property closed: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property console: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property crypto: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property defaultStatus: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property defaultstatus: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property devicePixelRatio: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property document: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property embeds: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property event: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property frameElement: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property frames: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property history: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property images: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property innerHeight: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property innerWidth: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property length: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property locationbar: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property menubar: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property name: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property navigator: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property offscreenBuffering: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onabort: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onbeforeunload: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onblur: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onchange: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onclick: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property ondblclick: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onerror: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onfocus: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onkeydown: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onkeypress: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onkeyup: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onload: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onmousedown: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onmousemove: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onmouseout: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onmouseover: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onmouseup: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onmousewheel: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onreset: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onresize: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onscroll: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onsearch: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onselect: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onsubmit: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property onunload: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property opener: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property outerHeight: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property outerWidth: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property pageXOffset: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property pageYOffset: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property personalbar: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property plugins: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property screen: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property screenLeft: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property screenTop: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property screenX: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property screenY: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property scrollbars: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property scrollX: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property scrollY: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property self: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property status: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property statusbar: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property toolbar: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property window: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property parent: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property addEventListener: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property alert: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property atob: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property blur: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property btoa: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property captureEvents: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property clearInterval: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property clearTimeout: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property close: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property confirm: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property constructor: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property eval: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property find: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property focus: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property getComputedStyle: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property getMatchedCSSRules: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property getSelection: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property moveBy: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property moveTo: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property open: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property print: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property prompt: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property releaseEvents: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property removeEventListener: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property resizeBy: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property resizeTo: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property scroll: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property scrollBy: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property scrollTo: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property setInterval: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property setTimeout: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
PASS: Unable to set property stop: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.
MAIN WINDOW: !!-- Test ended--!!
